+++ lib/reina.rb
@@ -1,0 +1,9 @@
+require 'bundler'
+Bundler.require
+
+require 'reina/config'
+require 'reina/app'
+require 'reina/controller'
+require 'reina/github_controller'
+require 'reina/server'
+require 'reina/version'
