+++ lib/reina/version.rb
@@ -1,0 +1,3 @@
+module Reina
+  VERSION = '0.1.0'.freeze
+end
